Move customer verification email into the CustomerEmailVerification subsystem
Relocate the verification email from the legacy includes/emails structure to
src/Internal/CustomerEmailVerification/Emails/CustomerVerifyEmail.php (PSR-4),
registered via the woocommerce_email_classes filter in the subsystem boot class,
mirroring the StockNotifications email pattern. Reverts the change to the shared
class-wc-emails.php so the branch no longer touches that legacy file.
